BigQuery/requirements.txt
google-cloud-bigquery~=3.4.2
